Refactor controllers and update DTO mappings
Refactored multiple controllers to use constructor injection, the `is null` pattern for null checks, and the `ToDto` method for entity-to-DTO conversion. Updated `CreatedAtAction` methods to correctly reference the `Id` of response DTOs. Modified `DeleteCustomer` and `DeleteUserAsync` methods to return `NoContent` instead of `Ok`.

Reordered properties in `NetworkDeviceDto` and `ServerDto` classes. Added new mapping classes for `Backup`, `Firewall`, `NetworkDevice`, and `Server` entities. Removed old migration files and updated the `UpdatedAt` property to be nullable in several entities. Removed the `Logo` property from the `Customer` entity and its associated migrations.

+++ ApiWebApp/Controllers/FirewallController.cs
@@ -1,72 +1,42 @@
 ﻿using ApiWebApp.DAL.Model;
 using ApiWebApp.Dto;
+using ApiWebApp.Mapping;
 using DAL.Data;
 using DAL.Models;
-using DAL.Repositories;
+
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+
 
 namespace ApiWebApp.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class FirewallController : ControllerBase
+    public class FirewallController(IRepository<Firewall> firewallRepository, IRepository<Customer> customerRepository) : ControllerBase
     {
-        private readonly IRepository<Firewall> _firewallRepository;
-        private readonly IRepository<Customer> _customerRepository;
-
-        public FirewallController(IRepository<Firewall> firewallRepository, IRepository<Customer> customerRepository)
-        {
-            _firewallRepository = firewallRepository ?? throw new ArgumentNullException(nameof(firewallRepository));
-            _customerRepository = customerRepository ?? throw new ArgumentNullException(nameof(customerRepository));
-        }
+        private readonly IRepository<Firewall> _firewallRepository = firewallRepository ?? throw new ArgumentNullException(nameof(firewallRepository));
+        private readonly IRepository<Customer> _customerRepository = customerRepository ?? throw new ArgumentNullException(nameof(customerRepository));
 
         [HttpGet]
         public async Task<ActionResult<IEnumerable<FirewallDto>>> GetFirewalls()
         {
             var firewalls = await _firewallRepository.GetAllAsync();
-            var firewallDtos = firewalls.Select(firewall => new FirewallDto
-            {
-                Id = firewall.Id,
-                Version = firewall.Version,
-                Model = firewall.Model,
-                SerialNumber = firewall.SerialNumber,
-                IpAddress = firewall.IpAddress,
-                MacAddress = firewall.MacAddress,
-                License = firewall.License,
-                CreatedAt = firewall.CreatedAt,
-                IsActive = firewall.IsActive,
-                CustomerId = firewall.CustomerId
-            }).ToList();
+            var firewallDtos = firewalls.Select(firewall => firewall.ToDto()).ToList();
 
             return Ok(firewallDtos);
         }
 
         [HttpGet("{id}")]
         public async Task<ActionResult<FirewallDto>> GetFirewall(Guid id)
-        {
+        { if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var firewall = await _firewallRepository.GetByIdAsync(id);
-            if (firewall == null)
+            if (firewall is null)
             {
                 return NotFound();
             }
-            var firewallDto = new FirewallDto
-            {
-                Id = firewall.Id,
-                Version = firewall.Version,
-                Model = firewall.Model,
-                SerialNumber = firewall.SerialNumber,
-                IpAddress = firewall.IpAddress,
-                MacAddress = firewall.MacAddress,
-                License = firewall.License,
-                CreatedAt = firewall.CreatedAt,
-                IsActive = firewall.IsActive,
-                CustomerId = firewall.CustomerId
-            };
-
+           var firewallDto = firewall.ToDto();
             return Ok(firewallDto);
         }
 
@@ -77,80 +47,22 @@ namespace ApiWebApp.Controllers
             {
                 return BadRequest(ModelState);
             }
-            // Validate CustomerId
+          
             var customer = await _customerRepository.GetByIdAsync(firewallDto.CustomerId);
-            if (customer == null)
+            if (customer is null)
             {
-                return BadRequest("Invalid CustomerId");
+                return BadRequest("Invalid Customer ID");
             }
 
-            var firewall = new Firewall
-            {
-                Version = firewallDto.Version,
-                Model = firewallDto.Model,
-                SerialNumber = firewallDto.SerialNumber,
-                IpAddress = firewallDto.IpAddress,
-                MacAddress = firewallDto.MacAddress,
-                License = firewallDto.License,
-                CreatedAt = DateTime.UtcNow,
-                UpdatedAt = firewallDto.UpdatedAt,
-                IsActive = firewallDto.IsActive,
-                CustomerId = firewallDto.CustomerId
-            };
-
+            var firewall = firewallDto.ToEntity();
             await _firewallRepository.AddAsync(firewall);
-
             var createdFirewall = await _firewallRepository.GetByIdAsync(firewall.Id);
-            var firewallResponseDto = new FirewallDto
-            {
-                Id = createdFirewall.Id,
-                Version = createdFirewall.Version,
-                Model = createdFirewall.Model,
-                SerialNumber = createdFirewall.SerialNumber,
-                IpAddress = createdFirewall.IpAddress,
-                MacAddress = createdFirewall.MacAddress,
-                License = createdFirewall.License,
-                CreatedAt = createdFirewall.CreatedAt,
-                IsActive = createdFirewall.IsActive,
-                CustomerId = createdFirewall.CustomerId
-            };
-
-            return CreatedAtAction(nameof(GetFirewall), new { id = createdFirewall.Id }, firewallResponseDto);
+            var backupResponseDto = createdFirewall.ToDto();
+            return CreatedAtAction(nameof(GetFirewall), new { id = backupResponseDto.Id }, backupResponseDto);
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutFirewall(Guid id, FirewallDto firewallDto)
-        {
-            if (id != firewallDto.Id)
-            {
-                return BadRequest("ID mismatch");
-            }
-
-            var firewall = await _firewallRepository.GetByIdAsync(id);
-            if (firewall == null)
-            {
-                return NotFound();
-            }
-
-            // Update firewall properties
-            firewall.Version = firewallDto.Version;
-            firewall.Model = firewallDto.Model;
-            firewall.SerialNumber = firewallDto.SerialNumber;
-            firewall.IpAddress = firewallDto.IpAddress;
-            firewall.MacAddress = firewallDto.MacAddress;
-            firewall.License = firewallDto.License;
-            firewall.UpdatedAt = DateTime.UtcNow; // Update UpdatedAt to current UTC time
-            firewall.IsActive = firewallDto.IsActive;
-            firewall.CustomerId = firewallDto.CustomerId;
-
-            // Call repository update method
-            await _firewallRepository.UpdateAsync(firewall);
-
-            return Ok(firewall);
-        }
-
-        [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteFirewall(Guid id)
+        public async Task<IActionResult> UpdateFirewall(Guid id, FirewallDto firewallDto)
         {
             if (!ModelState.IsValid)
             {
@@ -161,10 +73,25 @@ namespace ApiWebApp.Controllers
             {
                 return NotFound();
             }
+            firewallDto.UpdateEntity(firewall);
+            await _firewallRepository.UpdateAsync(firewall);
+            var updatedFirewall = firewall.ToDto();
+            return Ok(updatedFirewall);
+        }
+
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteFirewall(Guid id)
+        {
+            
+            var firewall = await _firewallRepository.GetByIdAsync(id);
+            if (firewall is null)
+            {
+                return NotFound();
+            }
 
             await _firewallRepository.DeleteAsync(id);
 
-            return Ok(ModelState);
+            return NoContent();
         }
     }
 }
